+++ src/components/common/Authentication/LogoutRequiredWrapper.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import { FC, useEffect } from 'react';
 import Loader from 'react-loader-spinner';
 
-import { HOME_URL } from '~/libs/const/urls';
+import { HOME_URL } from '~/libs/constants/urls';
 
 export const LoginRequiredWrapper: FC = ({ children }) => {
   const [session, loading] = useSession();
