+++ src/components/domain/Page/molecules/PagePreviewCard/PagePreviewCard.tsx
@@ -2,7 +2,7 @@ import { VFC } from 'react';
 
 import styled from 'styled-components';
 
-import { imagePath } from '~/libs/const/imagePath';
+import { imagePath } from '~/libs/constants/imagePath';
 import { BootstrapBreakpoints } from '~/libs/interfaces/variables';
 
 import { Page } from '~/domains/Page';
